requirements.txt
Flask==1.1.4
itsdangerous==1.1.0
MarkupSafe==1.1.1
Jinja2==2.11.3
click==7.1.2
Werkzeug==1.0.1
python-ldap==3.3.1
flask-login==0.5.0
sqlalchemy==1.3.24
flask-restful==1.1.1
